refactor: move tests out of newfeaturestest so that they run on upgrade-downgrade tests too
